Bug 15287: (follow-up) QA fix for position of user icon
This patch corrects the CSS for the user icon which appears when the
browser width is very narrow.

To test, apply the patch, rebuild CSS, and clear your browser cache if
necessary. View the OPAC with and without a user logged in and confirm
that the user icon at the top of the page looks correct at narrow
browser widths.

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>