KD-3172 - Slowness in acquisition
parcel.pl is very slow when there are many pending orders - one of our
libraries has 2500+ pending orders in one invoice, and the page load
takes a minute or so.

The workflow for acquisitions in this page is as follows:

- Take an item, scan in the barcode so it shows up in the
  pending orders table
- Click on the Receive-link in the table, which goes to orderreceive.pl,
  where you receive the order. The page then reloads parcel.pl

This patch changes the pending orders table to not load anything by default,
and makes it so the user has to search for pending orders in the "Filter"
entries.