Bug 21068: Remove NorwegianPatronDB related code
Bug 11401 introduced code to support Norwegian national library card.
This code is too specific to be part of Koha as it, it should be a
plugin instead.
Moreover nobody uses it, but a modified version (see comment 3).

RMaint: Rather than backporting the removal as is, I have created a
deprecation warning to highlight the codes upcoming removal.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>