Bug 21341: Style button on acquisitions existing record search with Bootstrap
This patch adds Bootstrap button styles to two links: View MARC and
Order.

Unrelated: A "modal-wide" class has been added to the View MARC modal to
give more room for the MARC view.

To test, apply the patch and go to Acquisitions -> Vendor -> Basket ->
Add to basket -> From an existing record.

In the record search results, confirm that the View MARC and Order links
are now Bootstrap buttons. Test the View MARC button and confirm that
the modal preview fills most of the screen.

Signed-off-by: Claire Gravely <claire.gravely@bsz-bw.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>