Bug 20809: (follow-up) Add missing JS to view log page
If the log viewer is being used to track patron-related data the
patron-related JavaScript file should be included.

To test, apply the patch and open a patron record. In the left-hand
sidebar choose Modification Log. Test the patron image edit/replace
process from the patron image in the sidebar and confirm that it works
correctly.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Claire Gravely <claire.gravely@bsz-bw.de>

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>