Bug 8630: (follow-up) Template corrections
This patch makes two corrections to templates: One to correct the case
of a variable, an error which I introduced when adding html filters.

The other removes a redundant image tag from the recent comments
template.

To test, apply the patch and log in to the OPAC as a user with
checkouts. Cover images on the "your summar" page should appear
correctly.

View the "recent comments" page and confirm that images still appear
correctly.

Signed-off-by: Aleisha Amohia <aleishaamohia@hotmail.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>