Bug 21505: Box around 'Additional fields' does not contain the fields
This patch modifies the subscription detail page markup to eliminate a
confusing bordered <legend> tag. It is replaced with an <hr> to
delineate the sections and an <h3> tag to serve as the heading.

To test, apply the patch and regenerate the staff client CSS.

 - Configure one or more additional subscription fields.
 - View the detail page for a subscription and confirm that the
   "Additional fields" section is clearly labeled.

Signed-off-by: Claire Gravely <claire.gravely@bsz-bw.de>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
(cherry picked from commit 311fe87b31b9a65974e6020df301408251659e79)
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>