Bug 21492: Show subscriptions count in the sidebar menu
This patch modifies the include file for the sidebar menu of
bibliographic views so that a count of subscriptions is shown on the
subscriptions tab.

To test you should have bibliographic record with one or more
subscriptions. View that bibliographic record and confirm that the
sidebar menu tab shows the count of subscriptions.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

https://bugs.koha-community.org/show_bug.cgi?id=21291

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>