KD 3354: Rules for automatic framework codes
Allow setting rules in a system preference, what framework code should
be automatically set for biblio, based on the MARC record fields.

(Filed as bug 21559 in Koha community)

Does not affect anything if the MarcToFrameworkcodeAutoconvert
system preference rules aren't used. The "Default" framework is
always overridden, if the rules match, but some other framework
chosen by you is always used.

Explanation of the rule format:
-------------------------------

The syspref is YAML, with a specific format:

- <field>
   <value>: <frameworkcode>

If the <field> in the record contains <value>, the <frameworkcode>
is used for the record.

<field> is one of:
"xxx" (field xxx in the MARC record)
"xxx/yy" (field xxx, character location yy)
"xxx/yy-zz" (field xxx, character locations yy-zz)
"xxx$v" (field xxx, subfield v)

Fields can be combined with plus-sign, then the field values have
plus-sign between them. The field-rules are checked in order from
top to bottom, and the first rule with a matching value is used.
If the (sub)field has been repeated in the MARC record, only the first
matching (sub)field is checked. If none of the rules match, the framework
code is not changed.

Example:

- 008/35-37 + 000/06:
   fin+a: BKSFIN
- 000/06:
   a: BKS
   c: MU
   e: MP

This would make record with 000/06 = "a" AND 008/35-37 = "fin"
use the BKSFIN framework, and any other record with 000/06 = "a" use
BKS, 000/06 = "c" would be MU, and 000/06 = "e" would give MP.