Bug 21568: Add more spans with classes around callnumbers in OPAC for additional styling
It would be nice if the callnumber portion of the callnumber + label was in a span,
and it also makes sense to wrap the entire callnumber string in a span as well.
This was started on bug 21340 but I have found more callnumbers in the search results that need spans.

Test Plan:
1) Apply this patch
2) Search for something in the OAPC with multiple results
3) View the HTML source
4) Note the new spans around the callnumbers, like: <span class="CallNumber"

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>