Bug 21493: (18.05.x) Remove incomplete icon style from serial issues tabs
This patch adds CSS to exclude serial issues history links from having
an icon background. There is no part of the icon sprite which is
designed for that purpose.

This patch also tweaks the positioning of a couple of icons.

To test, apply the patch and regenerate the OPAC CSS
(https://wiki.koha-community.org/wiki/Working_with_Bootstrap_OPAC_LESS_files).

- Open the detail page for a bibliographic record in the OPAC which has
  subscriptions.
- Click the "More details" link.
- Confirm that the "Brief history" and "Full history" links have no
  icon.
- Click the "Full history" link and do the same.
- On the normal, MARC, and ISBD views, confirm that the icons for those
  links look correct.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>